fix package names
Acquired groupid is `com.github.jyane`.
So change package names to `com.github.jyane` from `jp.jyane`.
