+++ pom.xml
@@ -2,7 +2,7 @@
   xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/maven-v4_0_0.xsd">
   <modelVersion>4.0.0</modelVersion>
 
-  <groupId>jp.jyane</groupId>
+  <groupId>com.github.jyane</groupId>
   <artifactId>grpc-java-api-checker</artifactId>
   <version>0.1-SNAPSHOT</version>
