+++ src/main/java/com/github/jyane/grpc/annotations/checkers/InternalChecker.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package jp.jyane.grpc.annotations.checkers;
+package com.github.jyane.grpc.annotations.checkers;
 
 import com.google.auto.service.AutoService;
 import com.google.errorprone.BugPattern;
@@ -23,45 +23,25 @@ import com.google.errorprone.BugPattern.SeverityLevel;
 import com.google.errorprone.bugpatterns.BugChecker;
 import com.google.errorprone.matchers.Description;
 import com.sun.source.tree.Tree;
-import java.util.Map.Entry;
-import java.util.Optional;
 import javax.lang.model.element.AnnotationMirror;
-import javax.lang.model.element.AnnotationValue;
-import javax.lang.model.element.ExecutableElement;
 
 @AutoService(BugChecker.class)
 @BugPattern(
-    name = "ExperimentalApi",
-    summary = "@ExperimentalApi should not be used in application code",
-    explanation = "@ExperimentalApi should not be used in application code",
+    name = "Internal",
+    summary = "@Internal should not be used in application code",
+    explanation = "@Internal should not be used in application code",
     severity = SeverityLevel.ERROR,
     linkType = LinkType.CUSTOM,
     link = "https://github.com/grpc/grpc-java"
 )
-public final class ExperimentalApiChecker extends AnnotationChecker {
+public final class InternalChecker extends AnnotationChecker {
 
-  public ExperimentalApiChecker() {
-    super("io.grpc", "io.grpc.ExperimentalApi");
-  }
-
-  private Optional<String> findLink(AnnotationMirror annotation) {
-    // Currently, @ExperimentalApi may have a link.
-    for (Entry<? extends ExecutableElement, ? extends AnnotationValue> entry : annotation
-        .getElementValues().entrySet()) {
-      return Optional.of(entry.getValue().toString());
-    }
-    return Optional.empty();
+  public InternalChecker() {
+    super("io.grpc", "io.grpc.Internal");
   }
 
   @Override
   protected Description describe(Tree tree, AnnotationMirror annotation) {
-    String link = findLink(annotation).orElse(this.linkUrl());
-    return Description.builder(
-        tree,
-        this.canonicalName(),
-        link,
-        this.defaultSeverity(),
-        this.message())
-        .build();
+    return describeMatch(tree);
   }
 }
